+++ webknossos/webknossos/dataset/_utils/pims_images.py
@@ -15,12 +15,14 @@ class PimsImages:
     def __init__(
         self,
         images: Union[str, "pims.FramesSequence", List[Union[str, PathLike]]],
+        channel: Optional[int],
         timepoint: Optional[int],
         swap_xy: bool,
         flip_x: bool,
         flip_y: bool,
         flip_z: bool,
         use_bioformats: bool,
+        is_segmentation: bool,
     ) -> None:
         """
         During initialization the pims objects are examined and configured to produce
@@ -45,6 +47,7 @@ class PimsImages:
         del images
 
         ## arguments as inner attributes
+        self._channel = channel
         self._timepoint = timepoint
         self._swap_xy = swap_xy
         self._flip_x = flip_x
@@ -120,6 +123,9 @@ class PimsImages:
                     assert "t" in images.axes
                     self._default_coords["t"] = timepoint
             else:
+                _allow_channels_first = not is_segmentation
+                if isinstance(images, (pims.ImageSequence, pims.ReaderSequence)):
+                    _allow_channels_first = False
                 if len(images.shape) == 2:
                     self._img_dims = "yx"
                     self._iter_dim = ""
@@ -130,7 +136,9 @@ class PimsImages:
                         self._img_dims = "yxc"
                         self._iter_dim = ""
                     elif images.shape[0] == 1 or (
-                        images.shape[0] == 3 and images.dtype == np.dtype("uint8")
+                        _allow_channels_first
+                        and images.shape[0] == 3
+                        and images.dtype == np.dtype("uint8")
                     ):
                         self._img_dims = "cyx"
                         self._iter_dim = ""
@@ -180,7 +188,12 @@ class PimsImages:
             )
 
         self._first_n_channels = None
-        if self.num_channels > 3:
+        if self._channel is not None:
+            assert (
+                self._channel < self.num_channels
+            ), f"Selected channel {self._channel} (0-indexed), but only {self.num_channels} channels are available."
+            self.num_channels = 1
+        elif self.num_channels > 3:
             warnings.warn(
                 f"Found more than 3 channels ({self.num_channels}), clamping to the first 3."
             )
@@ -229,7 +242,12 @@ class PimsImages:
                             self._original_images
                         )
                 else:
-                    images_context_manager = pims.open(self._original_images)
+                    try:
+                        images_context_manager = pims.open(self._original_images)
+                    except TypeError:
+                        images_context_manager = pims.ImageSequence(
+                            self._original_images
+                        )
 
             with images_context_manager as images:
                 if isinstance(images, pims.FramesSequenceND):
@@ -286,7 +304,9 @@ class PimsImages:
                         image_slice = image_slice.swapaxes(-1, -2)
 
                     if "c" in self._img_dims:
-                        if self._first_n_channels is not None:
+                        if self._channel is not None:
+                            image_slice = image_slice[self._channel : self._channel + 1]
+                        elif self._first_n_channels is not None:
                             image_slice = image_slice[: self._first_n_channels]
                         assert image_slice.shape[0] == self.num_channels, (
                             f"Image shape {image_slice.shape} does not fit to the number of channels "